another canvas
yet still not working

+++ src/problem3/Problem3.java
@@ -6,7 +6,7 @@ package problem3;
 
 public class Problem3 {
     public static void main(String[] args) {
-        int n = 13195; //issue with too large number
+        double n = 13195.0; //issue with too large number
         double y = 600851475143.0;
         LargestPrimeFactor euler3 = new LargestPrimeFactor();
